+++ src/containers/Blog/Blog.css
@@ -5,3 +5,27 @@
     width: 80%;
     margin: auto;
 }
+
+.Blog ul {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    width: 100%;
+    
+}
+
+
+.Blog li {
+display: inline-block;
+margin: 20px;
+}
+
+.Blog a {
+    text-decoration: none;
+    color: black;
+}
+
+.Blog a:hover,
+.Blog a:active{
+    color: orange
+}
